extract progress indicator comp and styling to a new folder
