+++ src/components/progress-indicator/_progress-indicator.scss
@@ -1,0 +1,37 @@
+.player .seek-bar {
+  &.ad-break {
+    cursor: initial;
+
+    .progress-bar .progress {
+      background-color: $ads-color;
+    }
+  }
+
+  &.live {
+    .progress-bar .progress {
+      background-color: $live-color;
+    }
+  }
+
+  .progress-bar {
+    .progress {
+      position: absolute;
+      top: 0;
+      left: 0;
+      height: 100%;
+      border-radius: inherit;
+      background-color: $primary-color;
+    }
+    .buffered {
+      position: absolute;
+      top: 0;
+      left: 0;
+      height: 100%;
+      border-radius: inherit;
+      background-color: rgba(255, 255, 255, 0.3);
+    }
+  }
+}
+
+
+
