+++ src/components/progress-indicator/index.js
@@ -1,0 +1,1 @@
+export {ProgressIndicator} from './progress-indicator';
