+++ src/styles/style.scss
@@ -21,6 +21,7 @@
 @import '../components/volume/volume';
 @import '../components/fullscreen/fullscreen';
 @import '../components/seekbar/seekbar';
+@import '../components/progress-indicator/progress-indicator';
 @import '../components/seekbar-preview/seekbar-preview';
 @import '../components/time-display/time-display';
 @import '../components/video-playing-title/video-playing-title';
